Frontend: Rename `App.tsx` to lowercase `app.tsx`
