+++ apps/frontend/src/index.tsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client';
 
 import '@/styles/index.css';
 
-import App from './App';
+import App from './app';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
